Build system changes for protocol autogen
Instead of adding protocol.pl as a build step, I opted just to add the
protocol-data files into the repository, and let the developer rerun the
script whenever necessary.  This is mostly to make the build not depend
on Perl, which conceivably could be missing e.g. for a Win32 build.