Simple non-player entity tracking
Adds the "mobs" and "pickups" modes for tracking the eponymous types of
entity, toggled with "m" and "p" respectively. Currently all mobs are
shown, regardless of their y-level. Nothing special is done on
mouse-over for now (and some rearchitecturing will be required to do
so). But this is a step closer to closing issue #4.